Move transparent transaction builder to an independent module.
